test java

+++ Java/TEST_TWO/src/App.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class App {
     public static void main(String[] args) throws Exception {
         Scanner input = new Scanner(System.in);
-        int age = input.nextInt();
+        double age = input.nextDouble();
 
         System.out.println("Hello, World! " + age);
         input.close();
